✅ Add unit tests for builder in various states
